Remove misleading detail about target endpoint
The configured solution to this step in the course expects you to provide the endpoint that you'd use to create a label. The instructions, however, indicate you should find the endpoint that is used to create a label _on an Issue_, which I interpreted as meaning the one used to _add_ a label to an Issue, and that was incorrect. But Issues don't factor into this at all, so I'm proposing a clarification no the instructions.